feat(cm): add insert in selector for campaign and lot, add triggers for tables to new functions, add new functions and minor fix data types

+++ dbmodel/cm/parent_schema/utils/fct/gw_trg_campaign_x_feature_validate_type.sql
@@ -42,7 +42,7 @@ BEGIN
         'SELECT c.%I_type
          FROM PARENT_SCHEMA.%I p 
          JOIN PARENT_SCHEMA.cat_%I c ON p.%Icat_id = c.id 
-         WHERE p.%I_id = $1',
+         WHERE p.%I_id = $1::integer',
         v_feature, v_feature, v_feature, v_feature, v_feature
     );
